combine imdb and text classification datasets
